requirements.txt
Flask
dotenv
google-generativeai